Support separate client ID and secret for ropg
Supporting a different client ID and secret allow for webview
login windows which support MFA and password verification with ropg.
Without a separate ID and secret one would be force to use a native
login window without MFA support for login, which doesn't support MFA or
the setup of an MFA device.