Fixes for mix matching web and ROPG
Set a fallback for _xcreds_oidc_username as it is used in the ROPG flow.

Additionally, force using ROPG secrets when available and when verifying
password. This change will allow mixing a webview for login and
verification, and ropg for password verification.

Things to note: the group claims in Okta should be setup for the login
application.